Try at ":" and fixes on strncpy
strncpy wasn't working, because there was no piece of code even using it.  The
colon command, which allows to create new definitions, is introduced.  Right
now it will just define a word which enters a subroutine saving lr and exists.
There is still some work to be done on it, but this minimal case seems to work.

The allocated memory to be executed had to be mmapped as using the heap (brk)
didn't return any memory which could be executed (even after touching it with
mprotect)