Change dictionary format (unreference name space)
Previously the name space was allocated using sbrk.  Nevertheless, this is the
application space, and is supposed to be provided linear.  The mmap anonymous
area could also be deployed but is being used already for compilation, with
execution permissions, which shouldn't be mixed up with simple R/W data.

Instead of allocating another mmap area, we allocate now a fixed amount in the
dictionary so the names are stored by value in the dictionary.  This leaves
our current primitive dictionary with the following structure:
 * Name Space => 32 bytes
 * Code Space =>  4 bytes (pointer to subrutine)

Note that DPANS94 specifies a maximum identifier length of 31 characters.