Refactor/simplify compile() mammoth
compile() was growing like a wild beast.  This commit factors out
compile_entry() and compile_exit(), which helps (again) reusing compile(), so
a new context (an IF statement, for instace) won't emit extra push/pop/bx
instructions that actually belong to the entry/exit section of a subroutine.

Yet another improvement has been storing the compilation pointer into a
setter/getter iface, so we avoid pushing/poping the compilation pointer
register.  It also narrows a bit the iface of the compilation module, which is
desirable, given how complex it was getting.