Compilation system cleanup: smarter assembling via macros
Up to now, every different instruction with different parameters needed to
have an own (non-exported) symbol inside compile.o.  This was done for
simplicity in the beginning but is obviously no way to go.

This is the first try to get rid of those.  The idea is to have macros to
generate the code for us.  These are still not optimal, as the loading of
the instruction encoding is done immediate by immediate (byte by byte) but is
with no doubt cleaner.

This commit replaces the helpers_push_* and helpers_pop_* with the right
macros.